python/requirements.txt
grpcio==1.12.0
protobuf==3.5.2.post1
click==6.7
